requirements.txt
# MediaPipe 수화 인식 프로젝트 의존성

# 컴퓨터 비전 & MediaPipe
opencv-python==4.8.1.78
mediapipe==0.10.8

# 딥러닝 프레임워크
tensorflow==2.15.0
# torch==2.1.0  # PyTorch를 사용하려면 주석 해제
# torchvision==0.16.0

# 데이터 처리
numpy==1.24.3
pandas==2.1.3
scikit-learn==1.3.2

# 음성 변환 (TTS)
pyttsx3==2.90
gTTS==2.4.0

# 시각화
matplotlib==3.8.2
seaborn==0.13.0

# 유틸리티
tqdm==4.66.1
pyyaml==6.0.1
